Sort and then binary search completion results
In a test where ~40000 completion results were returned from clang,
the autocomplete operation went from ~3 seconds to ~1 second when
binary searching for the range in the results that starts with the
completion string we're searching for.

The remaining 1 second is mostly clang doing its work so unclear
if anything can be done to improve that from the script side.