Better multithreading and much faster completions
Tweaked the options that are sent to Clang for faster completions. Made some of my tests go from 1 second to ~20 milliseconds when triggering completions.

Also implemented better multithreading so that the mainthread is blocked less. There might  be issues since it's not very well tested yet.

There's a new command ctrl+d,ctrl+w to manually warm up the cache.