Added replacement of ${home} ${project_path:} and ${folder:} tokens in the configuration values.
${home} is replaced with the value of the HOME environment variable.

${project_path:<file>} tries to find a file with the given name in all the registered project folders and
returns the first file found, or the original file name if none is found.
Example: ${project_path:project.sublime-project} tries to find a file named "project.sublime-project" relative
to the current project's folders. If none is found, it is replaced with "project.sublime-project".

${folder:<file>} is replaced with the dirname of the given path.
Example: ${folder:/path/to/file} is replaced with "/path/to".

Replacement is done sequencially, first all ${home} tokens, then all ${project_path:} are resolved, and then the
${folder:} are replaced.