Issue #39 fix. Make sure the settings for the actual view is used.
For example when previewing a file, it is loaded but not activated
so the view we want to parse is not the active view.

There's still an issue when starting Sublime Text 2 and a
previous session is restored. Unclear what, if anything, can
be done about it though, so the source code comment will have
to do for now:

 Views can apparently be window less, in most instances getting
 the active_window will be the right choice (for example when
 previewing a file), but the one instance this is incorrect
 is during Sublime Text 2 session restore. Apparently it's
 possible for views to be windowless then too and since it's
 possible that multiple windows are to be restored, the
 "wrong" one for this view might be the active one and thus
 ${project_path} will not be expanded correctly.

 This will have to remain a known documented issue unless
 someone can think of something that should be done plugin
 side to fix this.