Fix for issue #151
Moved everything after the lock inside the try block. This ensures
the unlock is called if anything bad happens. Also added
a default setting for "diagnostic_ignore_regex" in code as
apparently ST2 doesn't properly load settings files if the plugin
was ignored but now isn't and thus the defaults set in the settings
file wouldn't matter as they would all be the default None unless
a setting was provided in code.