Fix libclang.dylib loading on OS X
After pull request #218, SublimeClang stopped working on OS X because it
exposed an interesting issue: libclang.dylib was built with
'libclang.dylib' as install id and libcache.dylib links to that with
that id. If we first try to load libclang.dylib from the full path (#218)
instead of only the file name (before #218), then either dyld or ctypes
will mistakenly try to load /usr/lib/libclang.dylib when it's trying to
load libcache.dylib, my guess is certain cache stopped working because
path mismatch.

To make sure libcache.dylib is always loaded with correct
libclang.dylib, change the install id of libclang.dylib to
@loader_path/libclang.dylib, so that when loading libcache.dylib, we
always look for libclang.dylib in the same directory as libcache.dylib.